Extract replies action creators
